% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bonds_indices}
\alias{bonds_indices}
\title{Bond Information Dataset}
\format{
A data frame with 212 rows and 8 variables:
\describe{
\item{symbol}{bond Symbol aka:Ticker e.g.("2315", "4035" ..etc)}
\item{companyNameEN}{Full bond name in English e.g.("Sadara Sukuk", "Bahri Sukuk")}
\item{companyNameAR}{Full bond name in Arabic}
\item{companyName}{Bond name in English e.g.("Sadara Sukuk" "Bahri Sukuk")}
\item{market_type}{Market type to differentiate it from Stock Market}
\item{isin}{System code, usually not used}
\item{bond_type}{Bondy type. "S" for private, "G" for government}
}
}
\usage{
bonds_indices
}
\description{
A dataset containing the bond symbols and other attributes of over 73 bonds on Tadawul website.
The variables are as follows:
}
\keyword{datasets}
