% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{cach_me_com}
\alias{cach_me_com}
\title{Cashing the extracted data from the internet}
\usage{
cach_me_com(df, symbol)
}
\arguments{
\item{df}{extracted data}

\item{symbol}{the company symbol}
}
\description{
Cashing the extracted data from the internet
}
