% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{cach_me_index}
\alias{cach_me_index}
\title{Cashing the extracted data from the internet}
\usage{
cach_me_index(df, index_type)
}
\arguments{
\item{df}{extracted data}
}
\description{
Cashing the extracted data from the internet
}
