% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{check_cached_index}
\alias{check_cached_index}
\title{Cashing the extracted data from the internet}
\usage{
check_cached_index(start_date, end_date)
}
\arguments{
\item{start_date}{The start date of the query. Date is a string that needs to be in yyyy-mm-dd format}

\item{end_date}{The end date of the query. Date is a string that needs to be in yyyy-mm-dd format}
}
\value{
It returns a list that will include a cached data frame if exists
}
\description{
Cashing the extracted data from the internet
}
