% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{comp_info}
\alias{comp_info}
\title{Get company info in a list}
\usage{
comp_info(comp_symbol)
}
\arguments{
\item{comp_symbol}{Company Symbol.}
}
\value{
list with named elements for including all company info including its industry peers.
}
\description{
Get company info in a list
}
\examples{
comp_info(2010)
}
