% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date_elements}
\alias{date_elements}
\title{Date element extractor}
\usage{
date_elements(date_str)
}
\arguments{
\item{date_str}{Date is a string that needs to be in yyyy-mm-dd format}
}
\value{
list containing date elements and full date string.
}
\description{
Date element extractor
}
