% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df_to_xts}
\alias{df_to_xts}
\title{Converting Data Frame to xts}
\usage{
df_to_xts(x)
}
\arguments{
\item{x}{Data Frame to be converted to xts object}
}
\value{
returns an xts object
}
\description{
Converting Data Frame to xts
}
\examples{
df <- get_company_records("2020-01-01", "2020-12-31", 2222)
df_to_xts(df)
}
