% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_df}
\alias{format_df}
\title{Title}
\usage{
format_df(df, type = "index")
}
\arguments{
\item{df}{unformatted data frame}

\item{type}{specify type of data frame provided for formatting}
}
\value{
formatted data frame
}
\description{
Title
}
