% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financials.R
\name{get_available_xbrl_statements}
\alias{get_available_xbrl_statements}
\title{Get available XBRL statements}
\usage{
get_available_xbrl_statements(xbrl_statement, period)
}
\arguments{
\item{xbrl_statement}{xbrl_table produced by the \code{\link[tasi]{get_xbrl_table}}}

\item{period}{financial statement period which takes a form of (Qx or Annual)-Year. e.g "Q1-2021" or "Annual-2021"}
}
\value{
character vector of all available xbrl statements
}
\description{
Get available XBRL statements
}
\examples{
get_available_xbrl_statements(get_xbrl_table(2222), "Annual-2023")
}
