% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financials.R
\name{get_balance_sheet}
\alias{get_balance_sheet}
\title{Get balance sheet statement for specified company and reporting period}
\usage{
get_balance_sheet(company_symbol, period_type = "q")
}
\arguments{
\item{company_symbol}{Company Symbol number}

\item{period_type}{financial statement reporting period type which is either "q" for quarterly or "y" for annually}
}
\value{
formatted financial statement as tibble data frame.
}
\description{
Get balance sheet statement for specified company and reporting period
}
\examples{
get_balance_sheet(1010,"q")
}
