% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financials.R
\name{get_cash_flow}
\alias{get_cash_flow}
\title{Get cash flow statement for specified company and reporting period
#'}
\usage{
get_cash_flow(company_symbol, period_type = "q")
}
\arguments{
\item{company_symbol}{Company Symbol number}

\item{period_type}{financial statement reporting period type which is either "q" for quarterly or "y" for annually}
}
\value{
formatted financial statement as tibble data frame.
}
\description{
Get cash flow statement for specified company and reporting period
#'
}
\examples{
get_cash_flow(1010,"q")

}
