% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_company_records.R
\name{get_company_records}
\alias{get_company_records}
\title{Historical Data of a specified company for a specified period.}
\usage{
get_company_records(start_date, end_date, company_symbol, use_cache = TRUE)
}
\arguments{
\item{start_date}{The start date of the query. Date is a string that needs to be in yyyy-mm-dd format}

\item{end_date}{The end date of the query. Date is a string that needs to be in yyyy-mm-dd format}

\item{company_symbol}{Company Symbol number}

\item{use_cache}{logical operator for using cached data}
}
\value{
returns a data frame of company historical stock performance
}
\description{
Historical Data of a specified company for a specified period.
}
\examples{
#This will extract the historical stock records for ARAMCO for full year 2020
## Not run:
get_company_records("2020-01-01", "2020-12-31", 2222)
## End(Not run)
}
