% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/market_stats.R
\name{get_daily_market_stats}
\alias{get_daily_market_stats}
\title{Getting daily market daily report}
\usage{
get_daily_market_stats()
}
\value{
data frame containing all daily report for all companies listed in Saudi Stock Market
}
\description{
Getting daily market daily report
}
\examples{
get_daily_market_stats()
}
