% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financials.R
\name{get_fin_statement}
\alias{get_fin_statement}
\title{General Function to obtain Financial Statements (No XBRL parsing)}
\usage{
get_fin_statement(company_symbol, period_type = "y", statement)
}
\arguments{
\item{company_symbol}{Company Symbol number}

\item{period_type}{financial statement reporting period type which is either "q" for quarterly or "y" for annually}

\item{statement}{Financial Statement name which can be one of the three : "income_statement", "balance_sheet" or "cash_flow"}
}
\value{
formatted financial statement as tibble data frame.
}
\description{
General Function to obtain Financial Statements (No XBRL parsing)
}
