% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_industry.R
\name{get_health}
\alias{get_health}
\title{Health Index
get all records of Health Index for a specified period.}
\usage{
get_health(start_date, end_date, use_cache = TRUE)
}
\arguments{
\item{start_date}{The start date of the query. Date is a string that needs to be in yyyy-mm-dd format}

\item{end_date}{The end date of the query. Date is a string that needs to be in yyyy-mm-dd format}

\item{use_cache}{logical operator for using cached data}
}
\value{
Data frame of the Health Index
}
\description{
Health Index
get all records of Health Index for a specified period.
}
\examples{
get_health("2020-01-01", "2020-12-31")

}
