% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_index_records.R
\name{get_index_records}
\alias{get_index_records}
\title{TASI Index records
get all records of TASI for a specified period.}
\usage{
get_index_records(start_date, end_date, use_cache = TRUE)
}
\arguments{
\item{start_date}{The start date of the query. Date is a string that needs to be in yyyy-mm-dd format}

\item{end_date}{The end date of the query. Date is a string that needs to be in yyyy-mm-dd format}

\item{use_cache}{logical operator for using cached data}
}
\value{
Data frame of the TASI index
}
\description{
TASI Index records
get all records of TASI for a specified period.
}
\examples{
get_index_records("2020-01-01", "2020-12-31")

}
