% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news_stand.R
\name{get_news}
\alias{get_news}
\title{Getting daily market daily report
This function enables you to fetch all the news and public announcments made by the company specified by its symbol}
\usage{
get_news(symbol, language = "ar")
}
\arguments{
\item{symbol}{Valid company symbol. Company Symbols are usually 4 digit number with non-leading zero. for example 2222, 2010"}

\item{language}{You can specify the news lanaguage. It is either "ar" or "en". Default value is "ar" because large number of news are not translated in English.}
}
\value{
returns a data frame of company news and announcments
}
\description{
Getting daily market daily report
This function enables you to fetch all the news and public announcments made by the company specified by its symbol
}
