% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_industry.R
\name{get_real_estate}
\alias{get_real_estate}
\title{Real Estate Index
get all records of Real Estate Index for a specified period.}
\usage{
get_real_estate(start_date, end_date, use_cache = TRUE)
}
\arguments{
\item{start_date}{The start date of the query. Date is a string that needs to be in yyyy-mm-dd format}

\item{end_date}{The end date of the query. Date is a string that needs to be in yyyy-mm-dd format}

\item{use_cache}{logical operator for using cached data}
}
\value{
Data frame of the Real Estate Index
}
\description{
Real Estate Index
get all records of Real Estate Index for a specified period.
}
\examples{
get_real_estate("2020-01-01", "2020-12-31")

}
