% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financials.R
\name{get_statement_xbrl}
\alias{get_statement_xbrl}
\title{XBRL Financial Statement Extractor}
\usage{
get_statement_xbrl(company_symbol, period, statement_type = NULL)
}
\arguments{
\item{company_symbol}{Company Symbol number}

\item{period}{financial statement period which takes a form of (Qx or Annual)-Year. e.g "Q1-2021" or "Annual-2021"}

\item{statement_type}{specific string that match XBRL specifications for financial statements. This can be obtained by \code{\link[tasi]{get_available_xbrl_statements()}}. Defaults to NULL.}
}
\value{
Data Frame of type tibble of specified XBRL Financial Statement
}
\description{
XBRL Financial Statement Extractor
}
\examples{
company_symbol <- 1020
period <- "Annual-2019"
get_statement_xbrl(company_symbol, period,"StatementOfOtherComprehensiveIncomeBeforeTax")
}
