% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/financials.R
\name{get_xbrl_table}
\alias{get_xbrl_table}
\title{Getting XBRL financial statement links}
\usage{
get_xbrl_table(company_symbol)
}
\arguments{
\item{company_symbol}{Company Symbol number}
}
\value{
Data Frame that includes all links
}
\description{
Getting XBRL financial statement links
}
\examples{
get_xbrl_table(2222)
}
