% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getCompanyRecords.R
\name{getCompanyRecords}
\alias{getCompanyRecords}
\title{Historical Data of a specified company for a specified period.}
\usage{
getCompanyRecords(startDate, endDate, companySymbol)
}
\arguments{
\item{startDate}{The start date of the query. Date is a string that needs to be in yyyy-mm-dd format}

\item{endDate}{The end date of the query. Date is a string that needs to be in yyyy-mm-dd format}

\item{companySymbol}{Company Symbol number}
}
\value{

}
\description{
Historical Data of a specified company for a specified period.
}
\examples{
#This will extract the historical stock records for ARAMCO for full year 2020
getCompanyRecords("2020-01-01", "2020-12-31", 2222)

}
