% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getIndexRecords.R
\name{getIndexRecords}
\alias{getIndexRecords}
\title{TASI Index records
get all records of TASI for a specified period.}
\usage{
getIndexRecords(fromDate, toDate, use_cache = TRUE)
}
\arguments{
\item{fromDate}{The start date of the query. Date is a string that needs to be in yyyy-mm-dd format}

\item{toDate}{The end date of the query. Date is a string that needs to be in yyyy-mm-dd format}

\item{use_cache}{logical operator for using cached data}
}
\value{
Data frame of the TASI index
}
\description{
TASI Index records
get all records of TASI for a specified period.
}
\examples{
getIndexRecords("2020-01-01", "2020-12-31")

}
