% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{legacy_date}
\alias{legacy_date}
\title{Converting legacy date format into the new format}
\usage{
legacy_date(leg_date)
}
\arguments{
\item{leg_date}{string in date format 'yyyy-mm-dd'}
}
\value{
string in date format 'mm/dd/yyyy'
}
\description{
Converting legacy date format into the new format
}
