% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_industry.R
\name{list_industries}
\alias{list_industries}
\title{List all available industries}
\usage{
list_industries()
}
\value{
returns a character vector containing all industries names
}
\description{
List all available industries
}
\details{
This function is to be used in conjunction with \code{industry_func(industry)} function.
}
\examples{
get_banks <- industry_func(list_industries()[14])
# now we have a function called get_banks()
}
