% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{num_format}
\alias{num_format}
\title{Converting strings to cleaned up numeric values}
\usage{
num_format(num)
}
\arguments{
\item{num}{string number}
}
\value{
clean numeric value
}
\description{
Converting strings to cleaned up numeric values
}
