% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseURL.R
\name{parseURL}
\alias{parseURL}
\title{Parsing URL call
This function construct the URL string used to request data from TADAWUL website.}
\usage{
parseURL(
  startDate,
  endDate,
  comp_symbol,
  startIndex,
  endIndex,
  type,
  adjusted = FALSE
)
}
\arguments{
\item{startDate}{The start date of the query. Date is a string that needs to be in mm/dd/yyy format}

\item{endDate}{The end date of the query. Date is a string that needs to be in mm/dd/yyy format}

\item{comp_symbol}{Company Symbol number}

\item{startIndex}{start of data indexing}

\item{endIndex}{end of data indexing}

\item{type}{Type of URL cal. Choices are "company" or "index"}

\item{adjusted}{adjustment string variable. Options are "yes or "no"}
}
\value{
parsed url link for data
}
\description{
Parsing URL call
This function construct the URL string used to request data from TADAWUL website.
}
