% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/privateGetIndex.R
\name{privateGetIndex}
\alias{privateGetIndex}
\title{privateGetIndex}
\usage{
privateGetIndex(startDate, endDate, adjustPeriod)
}
\arguments{
\item{startDate}{The start date of the query. Date is a string that needs to be in yyyy-mm-dd format}

\item{endDate}{The end date of the query. Date is a string that needs to be in yyyy-mm-dd format}

\item{adjustPeriod}{This is logical term. If TRUE prices based on prior to adjustment.}
}
\value{
: data frame of the TASI index
}
\description{
privateGetIndex
}
