% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_data.R
\name{request_data}
\alias{request_data}
\title{Data Crawler}
\usage{
request_data(
  startDate,
  endDate,
  type,
  company_symbol = NULL,
  adjustPeriod = FALSE
)
}
\arguments{
\item{startDate}{The start date of the query. Date is a string that needs to be in yyyy-mm-dd format}

\item{endDate}{The end date of the query. Date is a string that needs to be in yyyy-mm-dd format}

\item{type}{this can only be "index" or "company" it determines how the data will be retrieved}

\item{company_symbol}{Company Symbol number}

\item{adjustPeriod}{This is logical term. If TRUE prices based on prior to adjustment.}
}
\value{
: data frame of the TASI index
}
\description{
Data Crawler
}
