% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caching.R
\name{setup_caching_folders}
\alias{setup_caching_folders}
\title{This function ensures that there is a folder to save the downloaded files}
\usage{
setup_caching_folders()
}
\description{
This function ensures that there is a folder to save the downloaded files
}
