% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{stock_indices}
\alias{stock_indices}
\title{Stocks Information Dataset}
\format{
A data frame with 212 rows and 8 variables:
\describe{
\item{symbol}{Company Symbol aka:Ticker e.g.("8220","4002" ..etc)}
\item{companyURL}{Company profile URL e.g.("/wps/portal/tadawul/markets/equities/indices/today/!ut/ ..")}
\item{sectorName}{Sector name to which companies belog e.g.("Bank","Energy" ..etc)}
\item{sectorURL}{Sector URL e.g.("/wps/portal/tadawul/markets/equities/indices/today/!...etc")}
\item{companyNameEN}{Full company name in English e.g.("Weqaya Takaful Insurance and Reinsurance Co.")}
\item{companyNameAR}{Full company name in Arabic e.g.("شركة وقاية للتأمين وإعادة التأمين التكافلي")}
\item{companyName}{Company name in English e.g.("Weqaya Takaful Insurance and Reinsurance Co.")}
\item{market_type}{Market type to differentiate it from Bond Market. Here, it is always "M" here}
\item{tradingNameEn}{Short company name in English e.g.("Weqaya Takaful")}
\item{tradingNameAr}{Short company name in Arabic e.g.("وقاية للتكافل")}
\item{isin}{System code, usually not used}
}
}
\usage{
stock_indices
}
\description{
A dataset containing the stock symbols and other attributes of over 200 companies on Tadawul website.
The variables are as follows:
}
\keyword{datasets}
