% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{validate_input}
\alias{validate_input}
\title{Validating input function}
\usage{
validate_input(start_date, end_date, company_symbol = NULL)
}
\arguments{
\item{start_date}{Date to be validated.}

\item{end_date}{Date to be validated.}

\item{company_symbol}{Company Symbol to be validated. This is omitted by Default.}
}
\value{
This Function returns error message if any of the checks failed.
}
\description{
Validating input function
}
