% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_adj_price}
\alias{add_adj_price}
\title{Add adjusted prices to dividens to an xts object}
\usage{
add_adj_price(x, symbol, start_date, end_date)
}
\arguments{
\item{x}{xts object (data frames will be converted to xts)}

\item{symbol}{the symbol of the company}

\item{start_date}{The start date of the query. Date is a string that needs to be in yyyy-mm-dd format}

\item{end_date}{The end date of the query. Date is a string that needs to be in yyyy-mm-dd format}
}
\value{
xts object with adjusted prices
}
\description{
Add adjusted prices to dividens to an xts object
}
\examples{
start_d <- "2022-01-01"
end_d <- "2023-01-01"
symbol <- 2222
get_company_records(start_date = start_d, end_date = end_d, company_symbol = symbol) |>
df_to_xts() |>
add_adj_price(symbol = symbol, start_date = start_d, end_date = end_d)

}
