% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_adj_price}
\alias{add_adj_price}
\title{Add adjusted prices to dividens to an xts object}
\usage{
add_adj_price(x, symbol)
}
\arguments{
\item{x}{xts object (data frames will be converted to xts)}

\item{symbol}{the symbol of the company}
}
\value{

}
\description{
Add adjusted prices to dividens to an xts object
}
