% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_company_records.R
\name{getSymbols}
\alias{getSymbols}
\title{Historical Data of a group of companies for specific period.}
\usage{
getSymbols(
  start_date,
  end_date,
  symbol_vector,
  tidy = FALSE,
  use_cache = TRUE,
  auto_assign = TRUE
)
}
\arguments{
\item{start_date}{The start date of the query. Date is a string that needs to be in yyyy-mm-dd format}

\item{end_date}{The end date of the query. Date is a string that needs to be in yyyy-mm-dd format}

\item{symbol_vector}{vector of company symbol numbers}

\item{tidy}{logical operator to return aggregated tibble}

\item{use_cache}{logical operator for using cached data}

\item{auto_assign}{logical operator to specify if results should be assigned to the parent enviornment or returned}
}
\value{
returns a data frame of company historical stock performance
}
\description{
Historical Data of a group of companies for specific period.
}
\examples{
#This will extract the historical stock records for ARAMCO for full year 2020
## Not run:
getSymbols("2020-01-01", "2020-12-31", c(2222, 2050))
## End(Not run)


}
