% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseURL.R
\name{parseURL}
\alias{parseURL}
\title{Parsing URL call
This function construct the URL string used to request data from TADAWUL website.}
\usage{
parseURL(p, fromDate, toDate, comSymbol = NULL, type, adjustment = FALSE)
}
\arguments{
\item{p}{pagination numeric variable}

\item{fromDate}{The start date of the query. Date is a string that needs to be in yyyy-mm-dd format}

\item{toDate}{The end date of the query. Date is a string that needs to be in yyyy-mm-dd format}

\item{comSymbol}{Company Symbol number}

\item{type}{Type of URL cal. Choices are "company" or "index"}

\item{adjustment}{adjustment string variable. Options are "yes or "no"}
}
\value{
parsed URL string
}
\description{
Parsing URL call
This function construct the URL string used to request data from TADAWUL website.
}
